requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.7
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
pip~=21.3.1
wheel~=0.37.1
setuptools~=60.2.0
django-filter~=23.2